+++ src/main/java/com/accbdd/complicated_bees/screen/widget/LibraryInfoWidget.java
@@ -1,7 +1,7 @@
 package com.accbdd.complicated_bees.screen.widget;
 
 import com.accbdd.complicated_bees.bees.mutation.Mutation;
-import com.accbdd.complicated_bees.config.Config;
+import com.accbdd.complicated_bees.config.ServerConfig;
 import com.accbdd.complicated_bees.screen.LibraryMenu;
 import com.accbdd.complicated_bees.util.GuiHelper;
 import net.minecraft.ChatFormatting;
@@ -101,7 +101,7 @@ public class LibraryInfoWidget extends AbstractScrollWidget {
                 0xFFFFFF,
                 0xFFFFFF,
                 Component.translatable("jei.complicated_bees.chance"),
-                Component.literal(String.format("%.0f%%", Math.min(100, (mutationWidget.possibleMutations.get(mutationWidget.selected).getChance() * 100) + (mutationWidget.selectedResearched ? Config.CONFIG.researchBonus.get() * 100 : 0)))));
+                Component.literal(String.format("%.0f%%", Math.min(100, (mutationWidget.possibleMutations.get(mutationWidget.selected).getChance() * 100) + (mutationWidget.selectedResearched ? ServerConfig.SERVER_CONFIG.researchBonus.get() * 100 : 0)))));
         nextY = GuiHelper.drawTopAlignedCenteredWrappedText(graphics,
                 infoPanelX,
                 nextY,